Don't delete remote branch with merge queues
When running `gh pr merge --delete-branch`, no longer delete the remote branch if the PR is merged via a merge queue.

Attempting to delete the remote branch with a PR that will be merged with a merge queue can cause the PR to be unexpectedly closed and the remote branch deleted before a successful merge.